ftp service

+++ Divar/Controllers/HomeController.cs
@@ -22,6 +22,12 @@ namespace Divar.Controllers
             var totalPages = (int)Math.Ceiling((double)totalAds / pageSize);
             var ads = await _adRepository.GetAllAdvertisementsAsync(pageNumber, pageSize, category, searchTerm);
 
+            foreach (var ad in ads)
+            {
+                var (imageUrl1, imageUrl2, imageUrl3) = _ftpService.DownloadImages(ad.Id);
+                ad.ImageUrl = imageUrl1 ?? ad.ImageUrl; // تنظیم اولین عکس به‌عنوان ImageUrl
+            }
+
             ViewBag.TotalPages = totalPages;
             ViewBag.CurrentPage = pageNumber;
             ViewBag.CurrentCategory = category;
@@ -32,6 +38,7 @@ namespace Divar.Controllers
 
 
 
+
         public async Task<IActionResult> Search(string searchTerm)
         {
             // Redirect to Index action with pageNumber set to 1
@@ -39,9 +46,6 @@ namespace Divar.Controllers
         }
 
 
-
-
-        // Show details
         public async Task<IActionResult> Detail(int id)
         {
             var ad = await _adRepository.GetAdvertisementByIdAsync(id);
@@ -49,13 +53,20 @@ namespace Divar.Controllers
             {
                 return NotFound();
             }
+
+            var (imageUrl1, imageUrl2, imageUrl3) = _ftpService.DownloadImages(id);
+
+            ViewBag.ImageUrl1 = imageUrl1;
+            ViewBag.ImageUrl2 = imageUrl2;
+            ViewBag.ImageUrl3 = imageUrl3;
+
             return View(ad);
         }
 
 
 
 
-        // To create Advertisement
+        //ایجاد اگهی
         [Authorize(Policy = "RequireHomeSelectCategory")]
         [HttpGet]
         public IActionResult Create()
@@ -78,39 +89,41 @@ namespace Divar.Controllers
             var userId = HttpContext.Session.GetString("UserId");
             if (ModelState.IsValid)
             {
+                advertisement.CustomUserId = userId;
+                advertisement.CreatedDate = DateTime.Now;
+                await _adRepository.AddAdvertisementAsync(advertisement);
+
                 // آپلود عکس‌ها به سرور FTP
                 if (imageFile != null && imageFile.Length > 0)
                 {
-                    var result = _ftpService.UploadImageToFtp(imageFile);
+                    var result = _ftpService.UploadImageToFtp(imageFile, advertisement.Id);
                     if (result)
                     {
-                        advertisement.ImageUrl = imageFile.FileName;
+                        advertisement.ImageUrl = $"ftp://127.0.0.1/advertisement_{advertisement.Id}/" + imageFile.FileName;
                     }
                 }
 
                 if (imageFile2 != null && imageFile2.Length > 0)
                 {
-                    var result = _ftpService.UploadImageToFtp(imageFile2);
+                    var result = _ftpService.UploadImageToFtp(imageFile2, advertisement.Id);
                     if (result)
                     {
-                        advertisement.ImageUrl2 = imageFile2.FileName;
+                        advertisement.ImageUrl2 = $"ftp://127.0.0.1/advertisement_{advertisement.Id}/" + imageFile2.FileName;
                     }
                 }
 
                 if (imageFile3 != null && imageFile3.Length > 0)
                 {
-                    var result = _ftpService.UploadImageToFtp(imageFile3);
+                    var result = _ftpService.UploadImageToFtp(imageFile3, advertisement.Id);
                     if (result)
                     {
-                        advertisement.ImageUrl3 = imageFile3.FileName;
+                        advertisement.ImageUrl3 = $"ftp://127.0.0.1/advertisement_{advertisement.Id}/" + imageFile3.FileName;
                     }
                 }
 
-                advertisement.CustomUserId = userId; // Use string for userId
-                advertisement.CreatedDate = DateTime.Now;
-                await _adRepository.AddAdvertisementAsync(advertisement);
+                await _adRepository.UpdateAdvertisementAsync(advertisement);
 
-                // Clear session after saving advertisement
+                // پاک کردن session بعد از ذخیره آگهی
                 HttpContext.Session.Remove("SelectedCategory");
 
                 return RedirectToAction("Index");
@@ -120,6 +133,9 @@ namespace Divar.Controllers
 
 
 
+
+
+        //انتخاب دسته بندی
         [Authorize(Policy = "RequireHomeSelectCategory")]
         [HttpGet]
         public IActionResult SelectCategory()
@@ -142,7 +158,7 @@ namespace Divar.Controllers
         // Edit Advertisement
         [Authorize(Policy = "RequireHomeEdit")]
         [HttpPost]
-        public async Task<IActionResult> Edit(int id, Advertisement updatedAdvertisement)
+        public async Task<IActionResult> Edit(int id, Advertisement updatedAdvertisement, IFormFile newImageFile1, IFormFile newImageFile2, IFormFile newImageFile3)
         {
             if (ModelState.IsValid)
             {
@@ -151,6 +167,11 @@ namespace Divar.Controllers
                 {
                     return NotFound();
                 }
+
+                var oldImageUrl1 = ad.ImageUrl;
+                var oldImageUrl2 = ad.ImageUrl2;
+                var oldImageUrl3 = ad.ImageUrl3;
+
                 ad.Title = updatedAdvertisement.Title;
                 ad.Content = updatedAdvertisement.Content;
                 ad.Price = updatedAdvertisement.Price;
@@ -166,12 +187,14 @@ namespace Divar.Controllers
                 ad.HomeAddress = updatedAdvertisement.HomeAddress;
                 ad.HomeSize = updatedAdvertisement.HomeSize;
 
+                // ویرایش تصاویر در سرور FTP
+                _ftpService.EditImages(ad.Id, newImageFile1, newImageFile2, newImageFile3, oldImageUrl1, oldImageUrl2, oldImageUrl3);
+
                 await _adRepository.UpdateAdvertisementAsync(ad);
                 return RedirectToAction(nameof(Index));
             }
             return View(updatedAdvertisement);
         }
-
 
         [Authorize(Policy = "RequireHomeEdit")]
         [HttpGet]
@@ -184,6 +207,8 @@ namespace Divar.Controllers
             }
             return View(ad);
         }
+
+
 
 
 
@@ -206,9 +231,21 @@ namespace Divar.Controllers
         [HttpPost, ActionName("Delete")]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            var ad = await _adRepository.GetAdvertisementByIdAsync(id);
+            if (ad == null)
+            {
+                return NotFound();
+            }
+
+            // حذف پوشه تصاویر مرتبط با آگهی از سرور FTP
+            _ftpService.DeleteFtpDirectory(ad.Id);
+
+            // حذف آگهی از پایگاه داده
             await _adRepository.DeleteAdvertisementAsync(id);
+
             return RedirectToAction(nameof(Index));
         }
+
 
 
         // User dashboard
